sandboxes/02-direct-otlp-ingest/app/requirements.txt
flask==3.1.*
opentelemetry-api==1.30.*
opentelemetry-sdk==1.30.*
opentelemetry-instrumentation-flask==0.51b*
opentelemetry-exporter-otlp-proto-http==1.30.*
gunicorn==23.*
